Added XML to Object converter
With this new conversion now all the numbers, booleans, hexadecimal,
and other values will be converted from string to respective value
type. So the "false" value will be false…

Also changed the method that check if the value has their respective
variable type. This its just the beginning of a restructure of the
library himself